improve: change breadcrumbs component
