+++ src/app/@breadcrumbs/default.tsx
@@ -1,0 +1,5 @@
+const Default = () => {
+  return <div></div>;
+};
+
+export default Default;
