--- src/components/Layout/CatalogLayout/components/Breadcrumbs/utils/getDefaultText.ts
@@ -1,5 +1,0 @@
-export const getDefaultText = (path: string) =>
-  ({
-    search: "Поиск",
-    catalog: "Каталог",
-  }[path]);
